backend/requirements.txt
agate==1.6.1
anyascii==0.3.1
asgiref==3.5.2
attrs==22.2.0
Babel==2.12.1
beautifulsoup4==4.11.1
betterproto==1.2.5
boto3==1.26.83
botocore==1.29.83
certifi==2022.9.24
cffi==1.14.6
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.3
distlib==0.3.6
Django==4.1.2
django-appconf==1.0.5
django-compressor==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-modelcluster==6.0
django-permissionedforms==0.1
django-taggit==3.1.0
django-treebeard==4.5.1
django-webpack-loader==1.8.0
djangorestframework==3.14.0
docopt==0.6.2
draftjs-exporter==2.1.7
et-xmlfile==1.1.0
filelock==3.8.0
grpclib==0.4.3
h2==4.1.0
hologram==0.0.15
hpack==4.0.0
html5lib==1.1
hyperframe==6.0.1
idna==2.9
importlib-metadata==6.0.0
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
json-rpc==1.14.0
jsonschema==3.1.1
l18n==2021.3
leather==0.3.4
Logbook==1.5.3
Markdown==3.4.1
MarkupSafe==2.1.2
mashumaro==3.3.1
minimal-snowplow-tracker==0.0.2
msgpack==1.0.4
multidict==6.0.4
mysql-connector-python==8.0.32
networkx==2.8.8
openpyxl==3.0.10
packaging==23.0
parsedatetime==2.6
pathspec==0.10.3
Pillow==9.3.0
pipreqs==0.4.11
platformdirs==2.5.2
protobuf==3.20.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2022.7
PyYAML==6.0
rcssmin==1.1.0
requests==2.23.0
rjsmin==1.2.0
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.3.1
stringcase==1.2.0
telepath==0.3
text-unidecode==1.3
typing_extensions==4.5.0
Unidecode==1.3.6
urllib3==1.25.11
virtualenv==20.16.5
wagtail==4.1.1
wagtail-blocks==0.8.1
wagtailfontawesome==1.2.1
webencodings==0.5.1
Werkzeug==1.0.1
whitenoise==6.2.0
Willow==1.4.1
yarg==0.1.9
zipp==3.15.0
